+++ apps/router-e2e/__e2e__/compiler/app/index.tsx
@@ -3,8 +3,16 @@ import { Text } from 'react-native';
 
 import { useBananas, useFruit } from './hooks/useFruit';
 
+let count = 0;
+
+const getCount = () => ++count;
+
 export default function Page() {
   const [isMounted, setIsMounted] = useState(false);
+
+  // If React Compiler works, we expect the update above to not
+  // update the count result from this invocation
+  const count = getCount();
 
   useBananas();
 
@@ -20,15 +28,11 @@ export default function Page() {
   return (
     <>
       <Text testID="test-anchor">Test</Text>
-      {isMounted && <CompilerSlots />}
+      {isMounted && count === 1 && <Child />}
     </>
   );
 }
 
-function CompilerSlots() {
-  const compilerSlots = eval('$');
-  if ('length' in compilerSlots) {
-    return <Text testID="react-compiler">{compilerSlots.length}</Text>;
-  }
-  return null;
+function Child() {
+  return <Text testID="react-compiler">2</Text>;
 }